requirements.txt
streamlit
joblib
numpy
pillow